Removento metodos de formatação de horas da Interface e passando provisoriamente para classe Routes

+++ src/main/java/com/ilegra/jt/lancamentodehoras/service/ActivityService.java
@@ -2,7 +2,6 @@ package com.ilegra.jt.lancamentodehoras.service;
 
 import com.ilegra.jt.lancamentodehoras.dao.ActivityDAO;
 import com.ilegra.jt.lancamentodehoras.pojo.Activity;
-import com.ilegra.jt.lancamentodehoras.pojo.Project;
 import com.ilegra.jt.lancamentodehoras.pojo.User;
 import com.ilegra.jt.lancamentodehoras.validators.ActivityValidator;
 import java.util.List;
@@ -16,20 +15,21 @@ public class ActivityService {
     
     public Optional<Activity> findById(long id){
         return new ActivityDAO().find(id);
-    }
+    }  
     
     public void update(User user, Activity activity){
-        if (ActivityValidator.isValidEdit(activity))
-            new ActivityDAO().update(user,activity);
+        if (ActivityValidator.isValidEdit(activity))new ActivityDAO().update(user,activity);
     }
+    
     public void delete(User user, Activity activity){
-        if(activity.getId()!= null)
-            new ActivityDAO().delete(user,activity);
+        if(activity.getId()!= null)new ActivityDAO().delete(user,activity);
     }
+    
     public Activity convertOptionalToActivity(Optional<Activity> optionalActivity){
         return (optionalActivity.isPresent()) ?  optionalActivity.get() : null;            
     }
-    public List<Activity> findByMonth(User user,Project project, Short month){
+    
+    public List<Activity> findByMonth(User user,Short month){
            return  new ActivityDAO().listByMonth(user,month);
     }
 }
